requirements.txt
pymongo==3.4.0astor==0.7.1astroid==1.4.7astropy==1.2.1pandas==0.24.2jsonschema==2.6.0xlrd==1.0.0urllib3==1.23matplotlib==2.0.0opencv-pythoncontextlib2==0.5.3PyQt5==5.12.2PyQt5-sip==4.19.17requests==2.19.1requests-oauthlib==1.2.0numpy==1.16.3pickleshare==0.7.5Pillow